+++ common/src/main/java/net/conczin/immersive_gateways/Client.java
@@ -1,7 +1,0 @@
-package net.conczin.immersive_gateways;
-
-public class Client {
-    public static void init() {
-        // nop
-    }
-}