+++ common/src/main/java/net/conczin/immersive_gateways/Common.java
@@ -9,10 +9,6 @@ public class Common {
     public static final String MOD_ID = "immersive_gateways";
     public static final Logger LOGGER = LogManager.getLogger();
 
-    public static void init() {
-        // nop
-    }
-
     public static ResourceLocation locate(String name) {
         return new ResourceLocation(MOD_ID, name);
     }