+++ fabric/src/main/java/net/conczin/immersive_gateways/fabric/ClientFabric.java
@@ -1,7 +1,6 @@
 package net.conczin.immersive_gateways.fabric;
 
 import net.conczin.immersive_gateways.BlockEntityTypes;
-import net.conczin.immersive_gateways.Client;
 import net.conczin.immersive_gateways.block.GatewayBlockEntityRenderer;
 import net.fabricmc.api.ClientModInitializer;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
@@ -10,8 +9,6 @@ import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
 public class ClientFabric implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
-        Client.init();
-
         BlockEntityRenderers.register(BlockEntityTypes.GATEWAY, GatewayBlockEntityRenderer::new);
     }
 }