+++ forge/src/main/java/net/conczin/immersive_gateways/forge/CommonForgeModBus.java
@@ -20,10 +20,6 @@ public class CommonForgeModBus {
         );
     }
 
-    public CommonForgeModBus() {
-        Common.init();
-    }
-
     @SubscribeEvent
     public static void register(RegisterEvent event) {
         registerHelper(event, BuiltInRegistries.ITEM, Items::registerItems);