+++ changelog.md
@@ -1,5 +1,6 @@
 # 0.0.2
 
+* Portals now generate their target structure rather than searching for it, reducing initial loading time
 * Generated structures are now less likely to be buried
 * Fixed Sus sand/gravels loot tables